moved all Tag use cases into separate controllers
