+++ src/controllers/GetTagController.tsx
@@ -26,16 +26,3 @@ export const getTags = async () => {
     return res.data;
 }
 
-export const createTag = async (tag_name: string) => {
-    const res = await supabase
-        .from("Tags")
-        .insert({ tag_name: tag_name })
-        .select()
-        .single();
-
-    if (res.error) {
-        throw res.error;
-    }
-
-    return res.data.tag_name;
-}
